去除formula模块使用drinkmodel，使用formula

+++ app/src/main/java/com/inno/coffee/ui/home/HomeDrinksLayout.kt
@@ -82,11 +82,11 @@ fun HomeDrinksLayout(
                 ) {
                     currentList.forEach { drinkModel ->
                         val enable =
-                            viewModel.enableMask(size > 0, checking, drinkModel.productId.toInt())
-                        val select = selected.intValue == drinkModel.productId.toInt()
+                            viewModel.enableMask(size > 0, checking, drinkModel.productId)
+                        val select = selected.intValue == drinkModel.productId
 
                         DrinkItem(model = drinkModel, enableMask = enable, selected = select) {
-                            selected.intValue = drinkModel.productId.toInt()
+                            selected.intValue = drinkModel.productId
                             viewModel.startMakeDrink(drinkModel, mainScreen, checking)
                         }
                     }
